build: widen the warning set to what the code already passes, and gate it
A sweep for C++ hygiene, done by measurement rather than by taste: enable
a flag against the whole tree, keep it only if the tree is already clean
or the handful it finds are worth fixing.

The result is a good sign about the existing code. Under -Wshadow,
-Wold-style-cast, -Wcast-align, -Wnon-virtual-dtor, -Woverloaded-virtual,
-Wnull-dereference, -Wdouble-promotion and -Wimplicit-fallthrough it
compiles with zero warnings as it stands, so those are free to turn on
and only ever catch a future regression.

-Wconversion and -Wsign-conversion are the two usually dismissed as
noisy, and they found the one thing worth finding: rng::uniform_int
returns int64 and its result was being assigned straight into a
vector<int> in the Hayashi-Yoshida bootstrap. Harmless today, because the
draw indexes a quantity that fits an int - and a silent truncation the
day it does not. Now an explicit cast with the reason next to it.

Four more were real but benign and are explicit now rather than implicit:
a size_t into an iterator difference in the event study, an int factor in
a size_t product in lead_lag, and three count-to-double conversions in
the bench printer. The test-side ones were loop counters typed int while
indexing containers; those became size_t, which removes the mismatch at
every use instead of casting at each index.

-Wfloat-equal is deliberately NOT enabled, and the reason is in the file:
it flags three exact comparisons that are correct. Mid prices come from
integer cents so they are exactly representable, and the question asked is
"did this change at all" - an epsilon there would swallow genuine small
moves. A warning that would make correct code worse is not worth having.

-Werror is now on in all five CI configure steps and off locally, which
is what cmake/warnings.cmake said the plan was when the scaffold was
written: "CI can promote to -Werror once the codebase fills in." At 10,828
lines and 256 tests it has. Local builds stay warning-only so a compiler
upgrade breaks the build that gates merges rather than the one someone is
debugging in.

Verified it gates: reintroducing the lead_lag narrowing fails the
-Werror build. 256 tests green, levelization unchanged.

+++ src/analytics/lead_lag.cpp
@@ -154,7 +154,8 @@ LeadLagResult CrossCorrelationEstimator::estimate() const {
                               config_.bootstrap_lag_halfwidth);
 
   std::mt19937 engine(config_.bootstrap_seed);
-  std::vector<double> boot_a(static_cast<std::size_t>(num_blocks) * block);
+  std::vector<double> boot_a(static_cast<std::size_t>(num_blocks) *
+                             static_cast<std::size_t>(block));
   std::vector<double> boot_b(boot_a.size());
   std::vector<double> leads;
   leads.reserve(static_cast<std::size_t>(config_.bootstrap_resamples));
